% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limbare.R
\name{limbare}
\alias{limbare}
\title{limbare}
\usage{
limbare(mixed.model, alpha, seg.Z, npsi, tol, max.iter)
}
\arguments{
\item{mixed.model}{linear mixed effects model of class "lme".}

\item{alpha}{percentage of data after trimming}

\item{seg.Z}{the segmented variable(s), i.e. the continuous
covariate(s) which have a piece-wise linear relationship with
the response variable. It is a formula with no response variable,
such as seg.Z=~x or seg.Z=~x1+x2.}

\item{npsi}{A named vector or list meaning the number (and not locations) of breakpoints to be estimated.}

\item{tol}{tolerance level}

\item{max.iter}{max number of iterations}
}
\description{
Description
A robust method to estimate breakpoints on longitudinal studies.
This method combined with least trimmed squared technique to
accommodate outliers in the dataset. Fits linear mixed models
with piece-wise relationships between response and one or more
covariates.
}
\details{
Details
}
\examples{
obs_data=sample_data

mixed.model=lme(y~x+duration, random = ~1|Subject_ID/Eye, data=obs_data, na.action = na.omit)
model=limbare(mixed.model, seg.Z = ~x, npsi=list(x=1), alpha=0.9, tol = 0.005, max.iter = 200)

model$psi
summary(model$model)
}
