% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.limbare.R
\name{plot.limbare}
\alias{plot.limbare}
\title{plot.limbare}
\usage{
\method{plot}{limbare}(
  model,
  seg.Z,
  intercept = 0,
  break.point = FALSE,
  break.point.CI = FALSE,
  ...
)
}
\arguments{
\item{model}{limbare model object}

\item{seg.Z}{the segmented variable(s), i.e. the continuous
covariate(s) which have a piece-wise linear relationship with
the response variable. It is a formula with no response variable,
such as seg.Z=~x.}

\item{intercept}{a constant to adjust the piece-wise lines vertically.
Default is 0.}

\item{...}{some settings for this generic require additional arguments.}

\item{breakpoint}{If TRUE, a vertical line to indicate the estimated
breakpoints will be presented. Default is FALSE.}

\item{breakpoint.CI}{If TRUE, two vertical lines to indicate the 95% CI
of the estimated breakpoints will be presented. Default is FALSE.}
}
\description{
Description
}
\details{
Details
}
\examples{
plot.limbare(model, seg.Z = ~x, intercept = 50, break.point = T,
             break.point.CI = T, color = "red")+
  geom_point(data=obs_data, aes(x, y), color = "red")+
  geom_point(data = model$data, aes(x, y))
}
