% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Simulated sample data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
Simulated
}
\usage{
data(sample_data)
}
\description{
Simulated data that contains 50 subjects and 60% of the subjects had
measurements from both eyes in which each eye had 3 or 4 repeated measurements
with 3% of outlier observations.
The true breakpoints were set at -4, and slopes 1, and 3.
}
\examples{
data(sample_data)

}
\references{
Simulation
}
\keyword{datasets}
