% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{slope}
\usage{
slope(model, seg.Z)
}
\arguments{
\item{model}{limbare model object}

\item{seg.Z}{the segmented variable(s), i.e. the continuous
covariate(s) which have a piece-wise linear relationship with
the response variable. It is a formula with no response variable,
such as seg.Z=~x.}
}
\description{
Description
Returns the estimated slopes and standard errors of each segment.
}
\details{
Details
}
\examples{
slope(model = model, seg.Z = ~x)
}
